Add an "interim" Level - 1.5 , and disable later levels for now.
With the low HP and Casts, moving from lvl 1 to lvl 2 (one archer vs a max of three archers...) was pretty rough :P. The interim level ("1.5") was added to smooth the transition/learning curve a little.

Besides a few other minor changes, levels are left unchanged otherwise.

(Minor changes being - reduce first wave of first level to 2 Villagers (from 4). And add a couple Villagers to the Champion in Level 2.)